# docs/api/openapi-specification.yaml
openapi: 3.0.0
info:
  title: Portfolio Site REST API
  version: 2.0.0
  description: |
    Comprehensive REST API for Hugo-based portfolio site with content management,
    analytics, webhooks, and GraphQL integration.
    
    ## Features
    - Full CRUD operations for all content types
    - Advanced search and filtering
    - Real-time analytics and reporting
    - Webhook system for integrations
    - GraphQL endpoint for flexible queries
    - Multi-language content support
    - Image optimization and management
    - SEO and metadata management
    
  contact:
    name: API Support
    email: support@portfolio.com
    url: https://portfolio.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.portfolio.com/v2
    description: Production server
  - url: https://staging-api.portfolio.com/v2
    description: Staging server
  - url: http://localhost:3000/api/v2
    description: Development server
security:
  - bearerAuth: []
  - apiKey: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and return JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: 'admin@portfolio.com'
              password: 'securePassword123'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

  # Content Management - Blog Posts
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: admin@portfolio.com
        password:
          type: string
          minLength: 8
          example: securePassword123
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: JWT refresh token
        expiresIn:
          type: integer
          description: Token expiration time in seconds
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, editor, viewer]
        firstName:
          type: string
        lastName:
          type: string
        avatar:
          type: string
          format: uri
        isActive:
          type: boolean
        preferences:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Authentication required'
            code: 'UNAUTHORIZED'
            timestamp: '2025-01-15T10:30:00Z'
...
    ValidationError:
      description: Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Validation failed'
            code: 'VALIDATION_ERROR'
            details:
              email: 'Invalid email format'
              password: 'Password must be at least 8 characters'
            timestamp: '2025-01-15T10:30:00Z'
...